fix errors of data_path

+++ data_path/request_pid_app.cpp
@@ -4,7 +4,7 @@
 #include <sys/shm.h>
 #include <fcntl.h>
 #include <unistd.h>
-#include <sys/syscall.h> 
+#include <sys/syscall.h>
 /* Not technically required, but needed on some UNIX distributions */
 #include <sys/types.h>
 #include <sys/stat.h>
@@ -27,7 +27,7 @@ int write(unsigned int fd, char*_buff, int size, int pid){
     char buff[100];
     strncpy(buff, _buff, size);
     strncpy(buff + size, (char*)&pid, 4);
-    return syscall(write, fd, buff, size + 4);
+    return syscall(write, fd, buff, (long int)(size + 4));
 }
 
 char*string = "hello world!";
